requirements.txt
mysql-connector-python==9.4.0
python-dateutil==2.9.0.post0
six==1.17.0
